% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weights.R
\name{add_weights}
\alias{add_weights}
\title{Add weights to a data frame for weighed estimates}
\usage{
add_weights(weighedEstimates, weightProfiles, weightProfileNames)
}
\arguments{
\item{weighedEstimates}{A `weighedEstimates` data frame as returned
by [mdmcda::build_weighed_estimate_df()].}

\item{weightProfiles}{The weight profiles object}

\item{weightProfileNames}{The name(s) of the weight profile(s) to add
weights for.}
}
\value{
The `weighedEstimates` object with added weights.
}
\description{
Add weights to a data frame for weighed estimates
}
