% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weights_to_criteriaTree.R
\name{add_weights_to_criteriaTree}
\alias{add_weights_to_criteriaTree}
\title{Add weights to criteria tree}
\usage{
add_weights_to_criteriaTree(
  weightsMeansAndSDs,
  criteria,
  weightCols,
  rootWeight = 1,
  rootParentCriterion_id = "-"
)
}
\arguments{
\item{weightsMeansAndSDs}{A data frame with weights.}

\item{criteria}{The criteria object.}

\item{weightCols}{The column names of the columns to add.}

\item{rootWeight}{The weight for the root (normally, 1).}

\item{rootParentCriterion_id}{The identifier of the root.}

\item{weightName}{The name to use for the variable with these weights.}
}
\value{
Invisibly, the `criteria` object.
}
\description{
This function takes weights from a data frame and adds
them to the criteria tree. NOTE: the
criteria tree is an R6 object. Therefore, when this function is called,
the object will be changed in memory; i.e. the `$criteriaTree` `data.tree`
object that is passed in `criteria` is passed by _reference_, not by value
(the default for normal R objects).
}
