% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_estimates_by_criterion.R
\name{aggregate_estimates_by_criterion}
\alias{aggregate_estimates_by_criterion}
\title{Aggregate estimates by criterion}
\usage{
aggregate_estimates_by_criterion(
  multiEstimateDf,
  estimateCol,
  criterionId_col = mdmcda::opts$get("criterionId_col"),
  fun = sum,
  ...
)
}
\arguments{
\item{multiEstimateDf}{A `multiEstimateDf` data frame.}

\item{estimateCol}{The column with the estimated to use.}

\item{criterionId_col}{The columns containing the criterion identifiers.}

\item{fun}{The function to use to aggregate the scores.}

\item{...}{Any additional arguments are passed to `fun`.}
}
\value{
A data frame with aggregated estimates.
}
\description{
Aggregate estimates by criterion
}
