% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_estimates_by_criterionCluster.R
\name{aggregate_estimates_by_criterionCluster}
\alias{aggregate_estimates_by_criterionCluster}
\title{Aggregate estimates by criterion clusters (parent criteria)}
\usage{
aggregate_estimates_by_criterionCluster(
  multiEstimateDf,
  estimateCol,
  parentCriterionId_col = mdmcda::opts$get("parentCriterionId_col"),
  fun = sum,
  ...
)
}
\arguments{
\item{multiEstimateDf}{A `multiEstimateDf` data frame.}

\item{estimateCol}{The column with the estimates to use.}

\item{parentCriterionId_col}{The columns containing the parent criterion
identifiers (i.e. the cluster identifiers).}

\item{fun}{The function to use to aggregate the scores.}

\item{...}{Any additional arguments are passed to `fun`.}
}
\value{
A data frame with aggregated estimates.
}
\description{
Aggregate estimates by criterion clusters (parent criteria)
}
