% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_estimates_by_decision.R
\name{aggregate_estimates_by_decision}
\alias{aggregate_estimates_by_decision}
\title{Aggregate the (potentially weighted) estimates by decision}
\usage{
aggregate_estimates_by_decision(
  multiEstimateDf,
  estimateCol,
  decisionId_col = mdmcda::opts$get("decisionId_col"),
  alternativeValue_col = mdmcda::opts$get("alternativeValue_col"),
  fun = sum,
  ...
)
}
\arguments{
\item{multiEstimateDf}{The `multiEstimateDf`.}

\item{estimateCol}{The name of the column with the estimates to aggregate.}

\item{decisionId_col, alternativeValue_col}{The columns containing the
decision identifiers and the values in each decision}

\item{fun}{The function to use to aggregate the estimates.}

\item{...}{Any additional arguments are passed to `fun`.}
}
\value{
A data frame with aggregated estimates.
}
\description{
Aggregate the (potentially weighted) estimates by decision
}
