% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_weights_and_criteria.R
\name{combine_weights_and_criteria}
\alias{combine_weights_and_criteria}
\title{Add weights from a data frame to criteria tree and update data frame}
\usage{
combine_weights_and_criteria(
  weightsMeansAndSDs,
  criteria,
  weightCols,
  rootWeight = 1,
  rootParentCriterion_id = "-"
)
}
\arguments{
\item{weightsMeansAndSDs}{The data frame with the weights; should have
a column `criterion_id` and one or more columns with weights.}

\item{criteria}{The `criteria` object.}

\item{weightCols}{A named character vector, where each value is a column
with weights in `weightsMeansAndSDs`, and each element's name is the
name that weight should get in the `criteria$criteriaTree` and in the
`weightsMeansAndSDs` that is returned after acculumation over the tree.}

\item{rootWeight}{The weight for the tree root.}

\item{rootParentCriterion_id}{The identifier for the tree root.}
}
\value{
The updated `weightsMeansAndSDs` (and `criteria$criteriaTree` is
updated in memory).
}
\description{
Add weights from a data frame to criteria tree and update data frame.
}
