% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores_per_alternative.R
\name{compute_scores_per_alternative}
\alias{compute_scores_per_alternative}
\title{Compute the scores per alternative}
\usage{
compute_scores_per_alternative(
  multiEstimateDf,
  weightProfiles,
  silent = FALSE,
  setMissingEstimates = 0,
  warnForMissingEstimates = !silent
)
}
\arguments{
\item{multiEstimateDf}{A `multiEstimateDf` that must contain the columns
specified in the `mdmcda` options `decisionId_col`, `alternativeValue_col`
(see [mdmcda::opts], and for each weight profile in `weightProfiles`, its
name appended with '`_weighted_estimate`'.}

\item{weightProfiles}{A `weightProfiles` object.}

\item{silent}{Whether to suppress messages.}

\item{setMissingEstimates}{What to set missing estimates to.}

\item{warnForMissingEstimates}{Whether to warn when encountering missing
estimates.}
}
\value{
A data frame.
}
\description{
Compute the scores per alternative
}
