% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_based_sensitivity_analysis.R
\name{confidence_based_sensitivity_analysis}
\alias{confidence_based_sensitivity_analysis}
\title{Conduct a sensitivity analysis based on confidence scores}
\usage{
confidence_based_sensitivity_analysis(
  multiEstimateDf,
  collapsedConfidences,
  scenarioDefinitions,
  weightProfiles,
  criteria,
  transformationFunction = setToZero,
  scenarioOrder = names(scenarioDefinitions),
  scenarioLabels = stats::setNames(names(scenarioDefinitions),
    names(scenarioDefinitions)),
  scorer = "all",
  confidenceThresholds = seq(0, 1, by = 0.1),
  setMissingEstimates = 0,
  silent = mdmcda::opts$get("silent"),
  lineSize = 1,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize"))
)
}
\arguments{
\item{multiEstimateDf}{A `multiEstimateDf`}

\item{theme}{}
}
\value{

}
\description{
When estimating the effects of alternatives on the criteria,
the raters can attach a confidence score to the estimates in the
performance table (or, more accurately, to the entire performance
table). This function uses these to test the robustness of the
results.
}
