% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labelDf.R
\name{create_labelDf}
\alias{create_labelDf}
\title{Create data frame with identifiers and labels}
\usage{
create_labelDf(namedVector, idCol, labelCol)
}
\arguments{
\item{namedVector}{The named vector, names should be identifiers,
values should be labels.}

\item{idCol, labelCol}{The names to use for the identifier column and the
label column.}
}
\value{
The data frame
}
\description{
Convenience function to convert a label vector to a label data frame.
}
\examples{
mdmcda::create_labelDf(
  c(
    id1 = "first element",
    id2 = "second element"
  ),
  idCol = "example_id",
  labelCol = "example_label"
);
}
