% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteriaCluster_df.R
\name{criteriaCluster_df}
\alias{criteriaCluster_df}
\title{Create a data frame with scores per criteria cluster per scenario}
\usage{
criteriaCluster_df(
  weighedEstimates,
  estimateCol,
  parentCriterionOrder = unique(weighedEstimates$parentCriterion_id),
  parentCriterionLabels = NULL,
  scenarioOrder = unique(weighedEstimates$scenario_id),
  scenarioLabels = NULL,
  sortByScore = FALSE,
  decreasing = TRUE
)
}
\arguments{
\item{weighedEstimates}{A `weighedEstimates` object.}

\item{estimateCol}{The column name with the estimates to use.}

\item{parentCriterionOrder}{The parent criteria to include.}

\item{parentCriterionLabels}{The labels for the parent criteria.}

\item{scenarioOrder}{The scenarios to include.}

\item{scenarioLabels}{The labels for the scenarios.}

\item{sortByScore, decreasing}{Whether to sort the scenarios by their total
scores, and if so, whether to sort them in decreasing order (from the left
side) or in increasing order.}
}
\value{
A data frame.
}
\description{
Create a data frame with scores per criteria cluster per scenario
}
