% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteriaCluster_df_perAlternative.R
\name{criteriaCluster_df_perAlternative}
\alias{criteriaCluster_df_perAlternative}
\title{Create a data frame with scores per criteria cluster per alternative}
\usage{
criteriaCluster_df_perAlternative(
  multiEstimateDf,
  estimateCol,
  criteria,
  alternativeLabels,
  parentCriterionOrder = unique(weightedEstimates$parentCriterion_id),
  parentCriterionLabels = NULL,
  decisionOrder = unique(weightedEstimates$decision_id),
  decisionLabels = NULL
)
}
\arguments{
\item{multiEstimateDf}{A `multiEstimateDf` data frame.}

\item{estimateCol}{The column name with the estimates to use.}

\item{parentCriterionOrder}{The parent criteria to include.}

\item{parentCriterionLabels}{The labels for the parent criteria.}

\item{decisionOrder}{The decisions to include.}

\item{decisionLabels}{The labels for the decisions.}
}
\value{
A data frame.
}
\description{
Create a data frame with scores per criteria cluster per alternative
}
