% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_scenarioDefinitions_to_xl.R
\name{export_scenarioDefinitions_to_xl}
\alias{export_scenarioDefinitions_to_xl}
\title{Export scenario definitions to an Excel spreadsheet}
\usage{
export_scenarioDefinitions_to_xl(
  scenarioDefinitions,
  file,
  scenarios = names(scenarioDefinitions)
)
}
\arguments{
\item{scenarioDefinitions}{The scenario definitions object.}

\item{file}{The file to write to.}

\item{scenarios}{Optionally, a selection of scenarios to export.}
}
\value{
Invisibly, the list of dataframes that's written to the file.
}
\description{
Scenario defintions are lists of vectors with selected alternatives.
Every vector is named, with the names being the decision identifiers of
each decision, and the values being the chosen alternative value for that
decision in that scenario.
}
