% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_weightProfile_to_xl.R
\name{export_weightProfile_to_xl}
\alias{export_weightProfile_to_xl}
\title{Export weight profiles to an Excel spreadsheet}
\usage{
export_weightProfile_to_xl(weightProfiles, file)
}
\arguments{
\item{weightProfiles}{The weightprofiles object, for example as created
by a call to [create_weight_profile()].}

\item{file}{The file to write to.}
}
\value{
Invisibly, the list of dataframes that's written to the file.
}
\description{
Weight profiles are lists of vectors with weights. Every vector is named,
with the names being the criterion identifiers of the criterion each weight
corresponds to. Every named vector (i.e. very weight profile) also has
an attribute that is another named vector: the weights of the criteria
clusters. Those are named vectors, too (again, with the names being the
criteria identifiers, but this time of the criteria clusters).
}
