% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alternativeLabel.R
\name{get_alternativeLabel}
\alias{get_alternativeLabel}
\title{Get the alternative labels based on a data frame}
\usage{
get_alternativeLabel(
  x,
  alternativeLabels,
  decisionId_col = mdmcda::opts$get("decisionId_col"),
  alternativeValue_col = mdmcda::opts$get("alternativeValue_col")
)
}
\arguments{
\item{x}{The data frame with a column with the decision identifier and the
alternative value.}

\item{alternativeLabels}{The object with the alternative labels (a named
list of named lists).}

\item{decisionId_colalternativeValue_col}{The names of the columns of the
decision identifier and the alternative value in data frame `x`.}
}
\value{
A vector with alternative labels for the alternative values for
each decision.
}
\description{
Get the alternative labels based on a data frame
}
