% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlighted_alternative_table.R
\name{highlighted_alternative_table}
\alias{highlighted_alternative_table}
\title{Create a scenario overview with highlighted alternatives}
\usage{
highlighted_alternative_table(
  scores_per_alternative,
  alternativeLabels,
  decreasing = FALSE,
  scenarioDefinition = NULL,
  decisionOrder = NULL,
  decisionLabels = NULL,
  colNames = c("Decisions and alternatives", "Scores"),
  caption = NULL,
  estimateParseFunction = NULL,
  omitAlternativeLabelRegex = NULL,
  returnTableOnly = TRUE,
  ...
)
}
\arguments{
\item{scores_per_alternative}{The `scorer_per_alternative` object as created
by [compute_scores_per_alternative()].}

\item{alternativeLabels}{The `alternativeLabels` object.}

\item{decreasing}{Whether to sort the alternatives in decreasing order
based on their final score.}

\item{scenarioDefinition}{The `scenarioDefinition` object to highlight;
if omitted, no highlighting is done.}

\item{decisionOrder}{The vector with the decision identifiers to include
(in the right order).}

\item{decisionLabels}{The named vector with the decision labels.}

\item{colNames}{The (two) column names to set in the resulting table.}

\item{caption}{The table caption to pass to [knitr::kable()].}

\item{estimateParseFunction}{Optionally, a function that is called to
parse the estimates before adding them to the table.}

\item{omitAlternativeLabelRegex}{A regular expression that can be used to
omit one or more alternatives based on their labels.}

\item{returnTableOnly}{Whether to return the table only, or the full
object that includes intermediate steps.}

\item{...}{Any additional arguments are passed on to `estimateParseFunction`.}
}
\description{
Create a scenario overview with highlighted alternatives
}
