% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scenarioDefinitions_from_xl.R
\name{import_scenarioDefinitions_from_xl}
\alias{import_scenarioDefinitions_from_xl}
\title{Import scenario definitions}
\usage{
import_scenarioDefinitions_from_xl(file)
}
\arguments{
\item{file}{The filename holding the scenario definitions as exported by
[export_scenarioDefinitions_to_xl()], where each worksheet has a
column `decision_id` and a column `alternative_value`.}
}
\value{
A list of vectors (i.e. a scenarioDefinitions object).
}
\description{
Import scenario definitions
}
