% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_if_knitting_to_html.R
\name{insert_if_knitting_to_html}
\alias{insert_if_knitting_to_html}
\title{Insert text only if knitting to HTML}
\usage{
insert_if_knitting_to_html(..., sep = "")
}
\arguments{
\item{...}{The text to insert; will be paste together the separator.}

\item{sep}{The separator.}
}
\value{
Nothing; the text is printed using [cat()].
}
\description{
Insert text only if knitting to HTML
}
\examples{
### This will show nothing unless you happen to knit this
### document, like in the Pkgdown version of the manual.
insert_if_knitting_to_html("This only shows up in the HTML version.");
}
