% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_estimates_from_xl.R
\name{load_estimates_from_xl}
\alias{load_estimates_from_xl}
\title{Load estimates from an Excel file}
\usage{
load_estimates_from_xl(file, multiEstimateDf = TRUE)
}
\arguments{
\item{file}{The filename.}

\item{multiEstimateDf}{Whether the file contains an exported
`multiEstimateDf`.}
}
\value{
The object with the estimates, containing the spreadsheet in
`$multiEstimateDf`.
}
\description{
This function loads estimates that have been exported
from an Excel file. For loading estimates from
completed performance tables, use [mdmcda::load_performance_tables()].
}
