% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\docType{data}
\name{opts}
\alias{opts}
\alias{set}
\alias{get}
\alias{reset}
\title{Options for the mdmcda package}
\format{
An object of class \code{list} of length 5.
}
\usage{
opts
}
\description{
The `mdmcda::opts` object contains three functions to set, get, and reset
options used by the dmcda package. Use `mdmcda::opts$set` to set options,
`mdmcda::opts$get` to get options, or `mdmcda::opts$reset` to reset specific or
all options to their default values.
}
\details{
It is normally not necessary to get or set `mdmcda` options.

The following arguments can be passed:

\describe{
  \item{...}{For `mdmcda::opts$set`, the dots can be used to specify the options
  to set, in the format `option = value`, for example,
  `varViewCols = c("values", "level")`. For
  `mdmcda::opts$reset`, a list of options to be reset can be passed.}
  \item{option}{For `mdmcda::opts$set`, the name of the option to set.}
  \item{default}{For `mdmcda::opts$get`, the default value to return if the
  option has not been manually specified.}
}

The following options can be set:

\describe{

  \item{decisionId_col}{The default column name for the decision identifier}
  \item{decisionLabel_col}{The default column name for the decision label}
  \item{criterionId_col}{The default column name for the criterion identifier}
  \item{criterionLabel_col}{The default column name for the criterion label}
  \item{criterionDescription_col}{The default column name for the criterion description}
  \item{parentCriterionId_col}{The default column name for the parent criterion identifier}
  \item{alternativeValue_col}{The default column name for the alternaive value}
  \item{alternativeLabel_col}{The default column name for the alternative label}
  \item{scenarioId_col}{The default column name for the scenario identifier}
  \item{scenarioLabel_col}{The default column name for the scenario label}
  \item{decisionDescription_col}{The default column name for the decision description}
  \item{decisionAlternatives_col}{The default column name for the label combining a decision and the selected alternative}
  \item{weightProfileId_col}{The default column name for the decision identifier}
  \item{score_col}{The column with the scores returned by `scenario_scores`}
  \item{leafCriterion_col}{The name of the column name indicating whether a criterion is a leaf (without children) or not}
  \item{rootCriterionId}{The name of the root criterion in the criaria tree}

}
}
\examples{
### Get the default columns in the variable view
mdmcda::opts$get(varViewCols);

### Set it to a custom version
mdmcda::opts$set(varViewCols = c("values", "level"));

### Check that it worked
mdmcda::opts$get(varViewCols);

### Reset this option to its default value
mdmcda::opts$reset(varViewCols);

### Check that the reset worked, too
mdmcda::opts$get(varViewCols);

}
\keyword{datasets}
