% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceSubtable_from_estimates.R
\name{performanceSubtable_from_estimates}
\alias{performanceSubtable_from_estimates}
\alias{all_performanceSubtables_from_estimates}
\title{Produce one or more 'clean' performance subtables}
\usage{
performanceSubtable_from_estimates(
  estimates,
  criteria,
  alternativeLabels,
  parentCriterion_id,
  criterionLabels,
  decision_id,
  decision_label,
  estimateColumn = "all",
  humanReadableOnly = FALSE
)

all_performanceSubtables_from_estimates(
  estimates,
  criteria,
  alternativeLabels,
  parentCriterionOrder,
  criterionLabels,
  decisionOrder,
  decisionLabels,
  estimateColumn = "all",
  humanReadableOnly = FALSE
)
}
\arguments{
\item{estimates}{The `estimates` object.}

\item{criteria}{The `criteri` object.}

\item{alternativeLabels}{The `alternativeLabels` object.}

\item{parentCriterion_id}{The identifier of the criterion cluster (parent
criterion) for which to produce the performance subtable.}

\item{criterionLabels}{The named vector with criterion (and parent
criterion) labels.}

\item{decision_id}{The identifier of the decisions for which to produce the
performance subtable.}

\item{decision_label}{The decision labels.}

\item{estimateColumn}{The column with the estimates; this can be changed to,
for example, the column with the estimates of a specific scorer, or the
weighted estimates, etc.}

\item{humanReadableOnly}{Whether to strip the columns with the decision
identifiers and the alternative values (if `humanReadableOnly=TRUE`).}

\item{parentCriterionOrder}{The vector with the parent criterion identifiers
to process (in the right order).}

\item{decisionOrder}{The vector with the decision identifiers to process (in
the right order).}

\item{decisionLabels}{The named vector with decision labels.}
}
\value{
For `performanceSubtable_from_estimates()`, a data frame; for
`all_performanceSubtables_from_estimates()`, a list of lists of data frames.
}
\description{
These functions take the `estimates`, `criteria`, and `alternativeLabels`
objects and produce one or more clean performance subtables.
}
