% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_criteria.R
\name{plot_criteria}
\alias{plot_criteria}
\title{Produce a DiagrammeR plot and show it and/or convert to SVG}
\usage{
plot_criteria(
  criteria,
  labels = NULL,
  wrapLabels = 60,
  color_nodes = TRUE,
  color_palette = "viridis",
  renderGraph = TRUE,
  returnSVG = FALSE,
  outputFile = NULL,
  ...
)
}
\arguments{
\item{criteria}{The criteria object as produced by a call to, for
example, [read_criteria_from_xl()]. The object should contain the
criteria tree in `$criteriaTree`.}

\item{labels}{Optionally, labels to apply to the criteria nodes.}

\item{wrapLabels}{If applying labels, the number of characters to wrap
them to.}

\item{color_nodes, color_palette}{Whether to use `DiagrammeR`s
`colorize_node_attrs()` function to color the nodes, and if so, which
palette to use.}

\item{renderGraph}{Whether to show the graph.}

\item{returnSVG}{Whether to return SVG or not. Valid values are `FALSE`,
to return the `DiagrammeR` object; `TRUE`, to return the SVG;
and anything that's not logical to return both in a list.}

\item{outputFile}{If not NULL and pointing to a file in an existing
directory, the graph will be exported to this file.}

\item{...}{Additional arguments are passed to `DiagrammeR`'s
`export_graph` function.}
}
\value{
Depending on the value of `returnSVG`, a `DiagrammeR` object;,
a character vector with SVG, or both in a list.
}
\description{
This renders a plot with the criteria tree. If the `criteriaTree` object
in the `criteria` object already contains the weights (as can be added with
[mdmcda::add_scorerWeights_to_criteriaTree()] or by a call
to [mdmcda::scorerWeights_to_profile()] or
[mdmda::combine_weights_and_criteria()]), those will be added and used to
color the tree and set the edge thicknesses.
}
