% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_confidences.R
\name{process_confidences}
\alias{process_confidences}
\title{Process the confidence scores and create plots}
\usage{
process_confidences(
  estimates,
  parentCriterionOrder = NULL,
  parentCriterionLabels = NULL,
  decisionOrder = NULL,
  decisionLabels = NULL,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize"))
)
}
\arguments{
\item{estimates}{The `estimates` object.}

\item{parentCriterionOrder, parentCriterionLabels}{The order and labels of
the parent criteria (i.e. the criterion clusters).}

\item{decisionOrder, decisionLabels}{The order and labels of the decisions.}

\item{theme}{The `ggplot2` theme to use for the plots.}
}
\value{
The `estimates` object, with plots and a data frame added.
}
\description{
Process the confidence scores and create plots
}
