% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scenarioDefinitions_in_columns_from_xl.R
\name{read_scenarioDefinitions_in_columns_from_xl}
\alias{read_scenarioDefinitions_in_columns_from_xl}
\title{Import scenario definitions defined in columns}
\usage{
read_scenarioDefinitions_in_columns_from_xl(
  file,
  decisionId_col = mdmcda::opts$get("decisionId_col"),
  decisionLabel_col = mdmcda::opts$get("decisionLabel_col"),
  decisionDescription_col = mdmcda::opts$get("decisionDescription_col"),
  decisionAlternatives_col = mdmcda::opts$get("decisionAlternatives_col")
)
}
\arguments{
\item{file}{The filename holding the scenario definitions, with the first
three columns specifying the decision identifiers, labels, and descriptions,
the fourth column specifying the concatenated alternative labels and values,
and the other columns specifying scenarios, with each column name being
a scenario identifier, and the values in each column indicating the selected
alternative value for each decision.}

\item{decisionId_col, decisionLabel_col, decisionDescription_col, decisionAlternatives_col}{The
column names.}
}
\value{
A list of vectors (i.e. a scenarioDefinitions object).
}
\description{
Import scenario definitions defined in columns
}
