% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weightsMeansAndSDs_from_xl.R
\name{read_weightsMeansAndSDs_from_xl}
\alias{read_weightsMeansAndSDs_from_xl}
\title{Read weightsMeansAndSDs data frame from an Excel file}
\usage{
read_weightsMeansAndSDs_from_xl(file)
}
\arguments{
\item{file}{The filename.}

\item{idCol, labelCol}{For `read_labels_from_xl()`, the name of the
columns with the identifiers and the labels.}
}
\value{
A data frame.
}
\description{
These function reads a weightsMeansAndSDs data frame from an Excel file.
}
