% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_alternative_table.R
\name{scenario_alternative_table}
\alias{scenario_alternative_table}
\title{Create a scenario overview with the selected alternatives}
\usage{
scenario_alternative_table(
  scenarioDefinition,
  alternativeLabels,
  decisionOrder = NULL,
  decisionLabels = NULL,
  colNames = c("Decision", "Alternative"),
  caption = NULL,
  returnTableOnly = TRUE
)
}
\arguments{
\item{scenarioDefinition}{The `scenarioDefinition` object.}

\item{alternativeLabels}{The `alternativeLabels` object.}

\item{decisionOrder}{The vector with the decision identifiers to include
(in the right order).}

\item{decisionLabels}{The named vector with the decision labels.}

\item{colNames}{The (two) column names to set in the resulting pretty table.}

\item{caption}{The table caption to pass to [knitr::kable()].}

\item{returnTableOnly}{Whether to return the table only, or the full
object that includes intermediate steps.}
}
\value{

}
\description{
Create a scenario overview with the selected alternatives
}
