% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_overview.R
\name{scenario_overview}
\alias{scenario_overview}
\title{Create an object with various useful data frames and plots for a scenario}
\usage{
scenario_overview(
  multiEstimateDf,
  estimateCol,
  scenario,
  scenarioLabel = NULL,
  createPlots = TRUE,
  decisionPlotOrder = "decreasing",
  criterionPlotOrder = "decreasing",
  criterionOrder = NULL,
  parentCriterionOrder = NULL,
  decisionOrder = NULL,
  criterionLabels = NULL,
  decisionLabels = NULL,
  alternativeLabels = NULL,
  parentCriterionIds_by_childId = NULL,
  verticalPlots = FALSE,
  useDecisionAlternativeLabels = TRUE,
  decision_alternative_pre = "**",
  decision_alternative_sep = "**:<br />",
  decision_alternative_suf = "",
  alternativeToOmitInTable = "Not applicable",
  scoreBarchart_criteria_args = NULL,
  scoreBarchart_decisions_args = NULL
)
}
\arguments{
\item{multiEstimateDf}{A `multiEstimateDf`.}

\item{estimateCol}{The column name with the estimates to use.}

\item{scenario}{The `scenarioDefinition` object with the scenario
definition: a named vector, with each element's name corresponding to
a decision, and each element's value to the alternative value for that
decision within the scenario.}

\item{createPlots}{Whether to create the plots.}

\item{decisionPlotOrder, criterionPlotOrder}{If no decisionOrder is
specified, whether to sort the decisions in the decision plot and criterion
plot in decreasing or increasing order.}

\item{criterionOrder, parentCriterionOrder, decisionOrder}{The order used to
organise the plots. These are vectors of identifiers.}

\item{criterionLabels, decisionLabels, alternativeLabels, scenarioLabel}{The
labels to use for the criteria, decisions, alternatives, and the scenario.
These are named vectors, with the elements being the labels, and their names
the corresponding identifiers (e.g. decision identifiers).}

\item{parentCriterionIds_by_childId}{Optionally, a named vector with the
names being the criteria, and the elements being their parent criterion. If
provided, the criteria plot will be organised and coloured by parent
criterion (i.e. criteria cluster).}

\item{verticalPlots}{Whether to print the plots vertically or horizontally.}

\item{useDecisionAlternativeLabels}{Whether to label the decision plot with
the decisions or combined labels that also include the alternative for
the scenario.}

\item{decision_alternative_pre, decision_alternative_sep, decision_alternative_suf}{If
`useDecisionAlternativeLabels` is `TRUE`, these prefix, separator, and suffix
are used to compose the decision plot labels.}

\item{alternativeToOmitInTable}{This alternative is not shown in the
table with highlighted options, unless it's selected for a scenario.}

\item{scoreBarchart_criteria_args, scoreBarchart_decisions_args}{These
arguments can be used to specify named lists with additional arguments for
`scoreBarChart_decisions` and `scoreBarChart_criteria`.}
}
\value{
And object with two plots and three data frames.
}
\description{
Create an object with various useful data frames and plots for a scenario
}
