% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreBarchart_criteria.R
\name{scoreBarchart_criteria}
\alias{scoreBarchart_criteria}
\title{Create a bar chart with scores per criterion}
\usage{
scoreBarchart_criteria(
  estimatesByCriterion,
  estimateCol,
  fill = "black",
  strokeColor = "black",
  strokeSize = 0.1,
  title = "MDMCDA criteria bar chart",
  xLab = "Criteria",
  yLab = estimateCol,
  criterionOrder = NULL,
  criterionLabels = NULL,
  parentCriterionOrder = NULL,
  parentCriterionIds_by_childId = NULL,
  xLabelRotationAngle = 45,
  verticalPlot = FALSE,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize")),
  guides = ggplot2::guide_legend(nrow = 2),
  legend.position = "top",
  legend.box.margin = ggplot2::margin(0.5, 0.5, 0.5, 0.5, "cm")
)
}
\arguments{
\item{estimatesByCriterion}{An `estimatesByCriterion` object as created by
[aggregate_estimates_by_criterion()].}

\item{estimateCol}{The column name with the estimates to use.}

\item{fill}{The color to use to fill the bars, if not by criterion cluster.}

\item{strokeColor, strokeSize}{The color and pen width of the stroke.}

\item{title, xLab, yLab}{The title and x and y axis labels.}

\item{criterionOrder}{The order of the criteria, a vector
with criterion identifiers.}

\item{criterionLabels}{The criterion labels, a names vector where every
element if a label and the corresponding name the criterion identifier.}

\item{parentCriterionOrder}{The order of the parent criteria (the clusters),
a vector with criterion identifiers.}

\item{parentCriterionIds_by_childId}{A names vector to easily find the
parent criterion identifier of each criterion identifier; the names should
be the criterion identifiers in `criterionOrder`, and the values the
corresponding parent criterion identifiers.}

\item{xLabelRotationAngle}{THe angel to rotate the x axis labels.}

\item{verticalPlot}{Whether to plot vertically.}

\item{theme}{The `ggplot2` theme to use.}

\item{guides}{A guides argument to tweak the legend.}

\item{legend.position, legend.box.margin}{The position and spacing for the
legend.}
}
\value{
A `ggplot2` plot.
}
\description{
Create a bar chart with scores per criterion
}
