% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreBarchart_criteria_for_singleDecision.R
\name{scoreBarchart_criteria_for_singleDecision}
\alias{scoreBarchart_criteria_for_singleDecision}
\title{Create a bar chart with scores per criterion for one alternative}
\usage{
scoreBarchart_criteria_for_singleDecision(
  multiEstimateDf,
  estimateCol,
  decision_id,
  scenario_id = NULL,
  scenarioDefinitions = NULL,
  alternative_value = NULL,
  fill = "black",
  strokeColor = "black",
  strokeSize = 0.1,
  title = "MDMCDA criteria bar chart for '\%s'",
  subtitle = "Scores for alternative '\%s' (total: \%s)",
  xLab = "Criteria",
  yLab = "Score",
  criterionOrder = NULL,
  criterionLabels = NULL,
  parentCriterionOrder = NULL,
  parentCriterionIds_by_childId = NULL,
  alternative_label = NULL,
  decisionLabels = NULL,
  alternativeLabels = NULL,
  xLabelRotationAngle = 45,
  normalizeScoreAxis = TRUE,
  verticalPlot = FALSE,
  scoreDigits = 0,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize")),
  guides = ggplot2::guide_legend(nrow = 2),
  legend.position = "top",
  legend.box.margin = ggplot2::margin(0.5, 0.5, 0.5, 0.5, "cm")
)
}
\arguments{
\item{multiEstimateDf}{A `multiEstimateDf` object.}

\item{estimateCol}{The column name with the estimates to use.}

\item{fill}{The color to use to fill the bars, if not by criterion cluster.}

\item{strokeColor, strokeSize}{The color and pen width of the stroke.}

\item{title, xLab, yLab}{The title and x and y axis labels.}

\item{criterionOrder}{The order of the criteria, a vector
with criterion identifiers.}

\item{criterionLabels}{The criterion labels, a names vector where every
element if a label and the corresponding name the criterion identifier.}

\item{parentCriterionOrder}{The order of the parent criteria (the clusters),
a vector with criterion identifiers.}

\item{parentCriterionIds_by_childId}{A names vector to easily find the
parent criterion identifier of each criterion identifier; the names should
be the criterion identifiers in `criterionOrder`, and the values the
corresponding parent criterion identifiers.}

\item{xLabelRotationAngle}{THe angel to rotate the x axis labels.}

\item{verticalPlot}{Whether to plot vertically.}

\item{theme}{The `ggplot2` theme to use.}

\item{guides}{A guides argument to tweak the legend.}

\item{legend.position, legend.box.margin}{The position and spacing for the
legend.}
}
\value{
A `ggplot2` plot.
}
\description{
Create a bar chart with scores per criterion for one alternative
}
