% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreBarchart_criteriaCluster_by_alternative.R
\name{scoreBarchart_criteriaCluster_by_alternative}
\alias{scoreBarchart_criteriaCluster_by_alternative}
\title{Create a bar chart with scores per criteria cluster for every alternative in a decision}
\usage{
scoreBarchart_criteriaCluster_by_alternative(
  multiEstimateDf,
  estimateCol,
  criteria,
  decision_id,
  alternativeValueRegex = ".*",
  parentCriterionOrder = unique(weightedEstimates$parentCriterion_id),
  parentCriterionLabels = parentCriterion_ids,
  decisionLabels = NULL,
  alternativeLabels = NULL,
  wrapLabels = 30,
  scoreRange = NULL,
  strokeColor = "black",
  strokeSize = 0.1,
  title = "MDMCDA criteria cluster bar chart by alternative",
  xLab = "Alternatives",
  yLab = estimateCol,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize")),
  guides = ggplot2::guide_legend(nrow = 1),
  legend.position = "top"
)
}
\arguments{
\item{multiEstimateDf}{A `multiEstimateDf` data frame.}

\item{estimateCol}{The column name with the estimates to use.}

\item{criteria}{A `criteria` object.}

\item{decision_id}{The identifier of the decision to show.}

\item{alternativeValueRegex}{Regex to select alternative values to include.}

\item{parentCriterionOrder}{The parent criteria to use, and their order.}

\item{decisionLabels}{The labels for the decisions (a named vector).}

\item{alternativeLabels}{The labels for the alternatives (a list of named vectors).}

\item{wrapLabels}{Where to wrap labels (`NULL` to not wrap).}

\item{strokeColor, strokeSize}{The color and pen width of the stroke.}

\item{title, xLab, yLab}{The title and x and y axis labels.}

\item{theme}{The `ggplot2` theme to use.}

\item{guides}{A guides argument to tweak the legend.}

\item{legend.position, legend.box.margin}{The position and spacing for the
legend.}

\item{parentCriterion_labels}{The labels for the parent criteria (a named vector).}
}
\value{
A `ggplot2` plot.
}
\description{
Create a bar chart with scores per criteria cluster for every alternative in a decision
}
