% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreBarchart_decisions.R
\name{scoreBarchart_decisions}
\alias{scoreBarchart_decisions}
\title{Create a bar chart with scores per decision}
\usage{
scoreBarchart_decisions(
  estimatesByDecision,
  estimateCol,
  fill = "black",
  strokeColor = "black",
  strokeSize = 0.1,
  title = "MDMCDA scores per decision",
  xLab = "Decisions",
  yLab = estimateCol,
  decisionOrder = NULL,
  decisionLabels = NULL,
  xLabelRotationAngle = 45,
  verticalPlot = FALSE,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize")),
  guides = ggplot2::guide_legend(nrow = 2),
  legend.position = "top",
  legend.box.margin = ggplot2::margin(0.5, 0.5, 0.5, 0.5, "cm")
)
}
\arguments{
\item{estimatesByDecision}{An `estimatesByDecision` object as created by
[aggregate_estimates_by_decision()].}

\item{estimateCol}{The column name with the estimates to use.}

\item{fill}{The color to use to fill the bars, if not by criterion cluster.}

\item{strokeColor, strokeSize}{The color and pen width of the stroke.}

\item{title, xLab, yLab}{The title and x and y axis labels.}

\item{decisionOrder}{The order of the decisions, a vector
with decision identifiers.}

\item{decisionLabels}{The decision labels, a names vector where every
element if a label and the corresponding name the decision identifier.}

\item{xLabelRotationAngle}{THe angel to rotate the x axis labels.}

\item{verticalPlot}{Whether to plot vertically.}

\item{theme}{The `ggplot2` theme to use.}

\item{guides}{A guides argument to tweak the legend.}

\item{legend.position, legend.box.margin}{The position and spacing for the
legend.}
}
\value{
A `ggplot2` plot.
}
\description{
Create a bar chart with scores per decision
}
