% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreBarchart_scenariosByCluster.R
\name{scoreBarchart_scenariosByCluster}
\alias{scoreBarchart_scenariosByCluster}
\title{Create a bar chart with scores per criteria cluster per scenario}
\usage{
scoreBarchart_scenariosByCluster(
  weightedEstimates,
  estimateCol,
  parentCriterionOrder = unique(weightedEstimates$parentCriterion_id),
  parentCriterionLabels = NULL,
  scenarioOrder = unique(weightedEstimates$scenario_id),
  scenarioLabels = NULL,
  sortByScore = FALSE,
  decreasing = TRUE,
  strokeSize = 0,
  strokeColor = "black",
  title = "MDMCDA scores by scenarios by criteria cluster",
  xLab = "Scenario",
  yLab = estimateCol,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize")),
  guides = ggplot2::guide_legend(nrow = 2),
  legend.position = "top"
)
}
\arguments{
\item{weightedEstimates}{A `weightedEstimates` object.}

\item{estimateCol}{The column name with the estimates to use.}

\item{scenarioOrder}{The scenarios to include.}

\item{scenarioLabels}{The labels for the scenarios.}

\item{sortByScore, decreasing}{Whether to sort the scenarios by their total
scores, and if so, whether to sort them in decreasing order (from the left
side) or in increasing order.}

\item{strokeColor, strokeSize}{The color and pen width of the stroke.}

\item{title, xLab, yLab}{The title and x and y axis labels.}

\item{theme}{The `ggplot2` theme to use.}

\item{guides}{A guides argument to tweak the legend.}

\item{legend.position, legend.box.margin}{The position and spacing for the
legend.}

\item{parentCriterion_ids}{The parent criteria to include.}

\item{parentCriterion_labels}{The labels for the parent criteria.}
}
\value{
A `ggplot2` plot.
}
\description{
Create a bar chart with scores per criteria cluster per scenario
}
