% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_postsets.R
\name{set_postsets}
\alias{set_postsets}
\title{Set postsets}
\usage{
set_postsets(
  multiEstimateDf,
  postsets,
  coder,
  decisionId_col = mdmcda::opts$get("decisionId_col"),
  criterionId_col = mdmcda::opts$get("criterionId_col"),
  alternativeValue_col = mdmcda::opts$get("alternativeValue_col")
)
}
\arguments{
\item{multiEstimateDf}{The `multiEstimateDf` to set the postsets in.}

\item{postsets}{The data frame with the postsets. This data frame must
have at least the four columns specified by the other four arguments.}

\item{coder}{The name of the column containing the estimates to replace
(typically estimated by a particular coder).}

\item{decisionId_col, criterionId_col, alternativeValue_col}{The names of
the columns with the decision and criterion identifiers and the alternative
value to which each estimate by `coder` pertains.}
}
\value{
The `multiEstimateDf` with replaced postset estimates.
}
\description{
Postsets are specifications of estimates in a non-performance-table
format that can be applied to a `multiEstimateDf`.
}
