% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weigh_multiEstimateDf.R
\name{weigh_multiEstimateDf}
\alias{weigh_multiEstimateDf}
\title{Weigh estimates in a multiEstimateDf}
\usage{
weigh_multiEstimateDf(
  multiEstimateDf,
  weightProfiles,
  scorer,
  weightProfileNames = names(weightProfiles)
)
}
\arguments{
\item{multiEstimateDf}{A multi estimate data frame that should contain
columns `decision_id`, `decision_label`, `alternative_value`,
`alternative_label`, `criterion_id`, `criterion_label`, and
one or more estimates in columns named with the scorer identifiers. Columns
with the `_id` suffix contain identifiers, and columns with the `_label`
suffix contain human-readable labels. This dataframe is stored in the
object called `multiEstimateDf` returned by a call to
[read_performance_tables()] to read a set of scored performance tables.}

\item{scorer}{The name of the scorer whose estimates to process.}

\item{weightProfileNames}{The weight profile name(s) to process.}

\item{weight_profiles}{The weight profiles: a named list of vectors, where
every vector element's value is the weight of a criterion, that
element's name is the criterion's identifier, and the vector's name is the
identifier of the weight profile.}
}
\value{
The `multiEstimateDf` with columns appended with the weights
and the weighted estimates.
}
\description{
Weigh estimates in a multiEstimateDf
}
