% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weigh_estimates_by_profile.R
\name{weight_estimates_by_profile}
\alias{weight_estimates_by_profile}
\title{Weigh estimates by one or more weight profiles}
\usage{
weight_estimates_by_profile(
  weighted_estimate_df,
  weight_profiles,
  weightProfileNames = names(weight_profiles)
)
}
\arguments{
\item{weighted_estimate_df}{A data frame with columns `scenario_id`,
`decision_id`, `alternative_value`, `criterion_id`, and `estimate`. This
data frame can be produced by a call to [weight_estimates_by_profile()].}

\item{weight_profiles}{The weight profiles: a named list of vectors, where
every vector element's value is the weight of a criterion, that
element's name is the criterion's identifier, and the vector's name is the
identifier of the weight profile.}

\item{weightProfileNames}{The weight profile name(s) to process.}
}
\value{
A data frame with columns `scenario_id`, `decision_id`,
`alternative_value`, `criterion_id`, `estimate`, and two columns for
every weight profile. These columns' names are the weight profile name
appended with `_weight` and `_weighted_estimate`.
}
\description{
Weigh estimates by one or more weight profiles
}
