% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_labels_to_xl.R
\name{write_labels_to_xl}
\alias{write_labels_to_xl}
\alias{write_criterionLabels_to_xl}
\alias{write_decisionLabels_to_xl}
\alias{write_scenarioLabels_to_xl}
\alias{write_alternativeLabels_to_xl}
\title{Write labels to an Excel file}
\usage{
write_labels_to_xl(
  x,
  file,
  idCol,
  labelCol,
  preventOverwriting = mdmcda::opts$get("preventOverwriting"),
  quiet = mdmcda::opts$get("quiet")
)

write_criterionLabels_to_xl(
  x,
  file,
  preventOverwriting = mdmcda::opts$get("preventOverwriting"),
  quiet = mdmcda::opts$get("quiet")
)

write_decisionLabels_to_xl(
  x,
  file,
  preventOverwriting = mdmcda::opts$get("preventOverwriting"),
  quiet = mdmcda::opts$get("quiet")
)

write_scenarioLabels_to_xl(
  x,
  file,
  preventOverwriting = mdmcda::opts$get("preventOverwriting"),
  quiet = mdmcda::opts$get("quiet")
)

write_alternativeLabels_to_xl(
  x,
  file,
  preventOverwriting = mdmcda::opts$get("preventOverwriting"),
  quiet = mdmcda::opts$get("quiet")
)
}
\arguments{
\item{x}{The object with the labels: a named vector, except for
alternative labels, which is a nested named list (see
[read_labels_from_xl()]).}

\item{file}{The filename to write to.}

\item{idCol, labelCol}{For `write_labels_to_xl()`, the name of the
columns with the identifiers and the labels.}
}
\value{
The written data frame.
}
\description{
These function writes labels to an Excel file (by first converting them
to a data frame).
}
\details{
The `write_labels_to_xl()` function is the underlying function; users
typically don't interact with this function, unless they specified custom
column names.
}
