% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_based_sensitivity_analysis.R
\name{confidence_based_sensitivity_analysis}
\alias{confidence_based_sensitivity_analysis}
\title{Conduct a sensitivity analysis based on confidence scores}
\usage{
confidence_based_sensitivity_analysis(
  multiEstimateDf,
  collapsedConfidences,
  scenarioDefinitions,
  weightProfiles,
  criteria,
  collapsedConfidences_criterionIdCol = "parentCriterion_id",
  transformationFunction = setToZero,
  scenarioOrder = names(scenarioDefinitions),
  scenarioLabels = stats::setNames(names(scenarioDefinitions),
    names(scenarioDefinitions)),
  scorer = "all",
  confidenceThresholds = seq(0, 1, by = 0.1),
  setMissingEstimates = 0,
  silent = mdmcda::opts$get("silent"),
  lineSize = 1,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize"))
)
}
\arguments{
\item{multiEstimateDf}{A multi estimate data frame that should contain
columns `decision_id`, `decision_label`, `alternative_value`,
`alternative_label`, `criterion_id`, `criterion_label`, and
one or more estimates in columns named with the scorer identifiers. Columns
with the `_id` suffix contain identifiers, and columns with the `_label`
suffix contain human-readable labels. This dataframe is stored in the
object called `multiEstimateDf` returned by a call to
[read_performance_tables()] to read a set of scored performance tables. Note
that different column names can be set using [mdmcda::opts].}

\item{collapsedConfidences}{The `collapsedConfidences` object.}

\item{scenarioDefinitions}{The `scenarioDefinitions` object.}

\item{weightProfiles}{The `weightProfiles` object.}

\item{criteria}{The `criteria` object.}

\item{collapsedConfidences_criterionIdCol}{The column name of the
criterion identifiers in the `collapsedConfidences` object to use.}

\item{transformationFunction}{The function to apply to the estimates in
performance tables below the confidence threshold}

\item{scenarioOrder}{The scenarios to process and the order in which to
organise them.}

\item{scenarioLabels}{The labels of the scenarios.}

\item{scorer}{The scorer of whom to process the confidences.}

\item{confidenceThresholds}{The confidence thresholds to process (by
default, 0 to 1 in steps of .1).}

\item{setMissingEstimates}{What to set missing values to, if any are
encountered.}

\item{silent}{Whether to suppress messages about progress.}

\item{lineSize}{The line size to use in the plots.}

\item{theme}{The `ggplot2` theme to use.}
}
\value{
An object with a data frame summarizing the sensitivity analysis,
and two produced plots: one with the scores, and one with the ranks.
}
\description{
When estimating the effects of alternatives on the criteria,
the raters can attach a confidence score to the estimates in the
performance table (or, more accurately, to the entire performance
table). This function uses these to test the robustness of the
results.
}
\details{
In the confidence-based sensitivity analyses, by default 10% of the
estimates about which the raters expressed the least confidence
are sequentially set to the result of the `transformationFunction`.
The total score for each scenario is then computed again.
}
