% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labelDf.R
\name{create_labelDf}
\alias{create_labelDf}
\title{Create data frame with identifiers and labels}
\usage{
create_labelDf(namedVector, type)
}
\arguments{
\item{namedVector}{The named vector, names should be identifiers,
values should be labels.}

\item{type}{The type is prefixed to the column names (the suffixes
`_id` and `_label` are added). Common types are "`decision`", "`criterion`",
and "`scenario`".}
}
\value{
The data frame
}
\description{
Convenience function to convert a label vector to a label data frame.
}
\examples{
create_labelDf(
  c(
    id1 = "first element",
    id2 = "second element"
  ),
  type = "example"
);
}
