% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\docType{data}
\name{opts}
\alias{opts}
\alias{set}
\alias{get}
\alias{reset}
\title{Options for the dmcda package}
\format{An object of class \code{list} of length 5.}
\usage{
opts
}
\description{
The `dmcda::opts` object contains three functions to set, get, and reset
options used by the dmcda package. Use `dmcda::opts$set` to set options,
`dmcda::opts$get` to get options, or `dmcda::opts$reset` to reset specific or
all options to their default values.
}
\details{
It is normally not necessary to get or set `dmcda` options.

The following arguments can be passed:

\describe{
  \item{...}{For `dmcda::opts$set`, the dots can be used to specify the options
  to set, in the format `option = value`, for example,
  `varViewCols = c("values", "level")`. For
  `dmcda::opts$reset`, a list of options to be reset can be passed.}
  \item{option}{For `dmcda::opts$set`, the name of the option to set.}
  \item{default}{For `dmcda::opts$get`, the default value to return if the
  option has not been manually specified.}
}

The following options can be set:

\describe{

  \item{varViewCols}{The order and names of the columns to include in the
  variable view.}

  \item{showLabellerWarning}{Whether to show a warning if labeller labels
  are encountered.}

}
}
\examples{
### Get the default columns in the variable view
dmcda::opts$get(varViewCols);

### Set it to a custom version
dmcda::opts$set(varViewCols = c("values", "level"));

### Check that it worked
dmcda::opts$get(varViewCols);

### Reset this option to its default value
dmcda::opts$reset(varViewCols);

### Check that the reset worked, too
dmcda::opts$get(varViewCols);

}
\keyword{datasets}
