% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceTable_heatmap.R
\name{performanceTable_heatmap}
\alias{performanceTable_heatmap}
\title{Create a performance table heatmap}
\usage{
performanceTable_heatmap(
  weightedEstimates,
  estimateCol,
  scenario_id,
  scenario_label = scenario_id,
  criterionOrder = NULL,
  decisionOrder = NULL,
  criterionLabels = NULL,
  decisionLabels = NULL,
  alternativeLabels = NULL,
  useDecisionAlternativeLabels = TRUE,
  decision_alternative_pre = "**",
  decision_alternative_sep = "**:<br />",
  decision_alternative_suf = "",
  criterionId_col = mdmcda::opts$get("criterionId_col"),
  parentCriterionId_col = mdmcda::opts$get("parentCriterionId_col"),
  criterionLabel_col = mdmcda::opts$get("criterionLabel_col"),
  decisionId_col = mdmcda::opts$get("decisionId_col"),
  decisionLabel_col = mdmcda::opts$get("decisionLabel_col"),
  alternativeValue_col = mdmcda::opts$get("alternativeValue_col"),
  alternativeLabel_col = mdmcda::opts$get("alternativeLabel_col"),
  scenarioId_col = mdmcda::opts$get("scenarioId_col"),
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize"))
)
}
\arguments{
\item{weightedEstimates}{A `weightedEstimates` data frame as produced by
[mdmcda::weight_estimates()].}

\item{estimateCol}{The name of the column with the estimates.}

\item{scenario_id, scenario_label}{The identifier and label of the scenario
to create the heatmap for.}

\item{criterionOrder, decisionOrder}{Optionally, vectors with criterion
and decision identifiers to indicate the order that should be used.}

\item{criterionLabels, decisionLabels}{Optionally, named vectors with labels
for the criteria and decisions. These have to be named vectors, with the
elements being the labels, and the names the identifiers.}

\item{alternativeLabels}{The `alternativeLabels` object; optionally, to
override the alternative labels in the `weightedEstimates` object.}

\item{useDecisionAlternativeLabels}{Whether to label the decision plot with
the decisions or combined labels that also include the alternative for
the scenario.}

\item{decision_alternative_pre, decision_alternative_sep, decision_alternative_suf}{If
`useDecisionAlternativeLabels` is `TRUE`, these prefix, separator, and suffix
are used to compose the decision plot labels.}

\item{criterionId_col, parentCriterionId_col, criterionLabel_col, decisionId_col, decisionLabel_col, alternativeValue_col, alternativeLabel_col, alternativeLabel_col, scenarioId_col}{The
column names in the `weightedEstimates` data frame.}

\item{theme}{The `ggplot2` theme to use.}
}
\value{
A `ggplot2` plot.
}
\description{
This is a convenient visualisation that shows the (weighted) estimates
for each decision/criterion combination. Decisions are sorted
alphabetically and criteria are sorted alphabetically within their
clusters, which are also sorted alphabetically.
}
