% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceTable_heatmap.R
\name{performanceTable_heatmap}
\alias{performanceTable_heatmap}
\title{Create a performance table heatmap}
\usage{
performanceTable_heatmap(
  weighedEstimates,
  estimateCol,
  scenario_id,
  criterionLabels = NULL,
  decisionLabels = NULL,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize"))
)
}
\arguments{
\item{weighedEstimates}{A `weighedEstimates` data frame as produced by
[mdmcda::weigh_estimates()].}

\item{estimateCol}{The name of the column with the estimates.}

\item{scenario_id}{The identifier of the scenario to create the heatmap
for.}

\item{criterionLabels, decisionLabels}{Optionally, named vectors with labels
for the criteria and decisions. These have to be named vectors, with the
elements being the labels, and the names the identifiers.}

\item{theme}{The `ggplot2` theme to use.}
}
\value{
A `ggplot2` plot.
}
\description{
This is a convenient visualisation that shows the (weighed) estimates
for each decision/criterion combination. Decisions are sorted
alphabetically and criteria are sorted alphabetically within their
clusters, which are also sorted alphabetically.
}
